# backend/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SysMap Social Network - OpenAPI 3.0",
    "description": "Backend of the Social Network project of SysMap fullstack bootcamp.",
    "contact": {
      "email": "kellyplcastelo@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000/v1"
    }
  ],
...
  "paths": {
...
    "/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Finds logged user",
        "description": "Finds logged user",
        "operationId": "findLoggedUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Updates logged user's username",
        "description": "Updates username",
        "operationId": "updateUser",
        "requestBody": {
          "description": "Updates logged in username",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "203": {
            "description": "UNDOCUMENTED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Deletes logged user",
        "description": "deletes logged user",
        "operationId": "deleteUser",
        "responses": {
          "204": {
            "description": "NO CONTENT"
          },
          "401": {
            "description": "UNAUTHORIZED"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "637bb9d4da4e28ea8606b7bf"
          },
          "user": {
            "type": "string",
            "example": "Castelo"
          },
          "profile": {
            "type": "string",
            "example": "637bb9d4da4e28ea8606b7c1"
          }
        }
      },
      "UserUpdateRequest": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "example": "Castelo"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}